+++ scripts/source_code.py
@@ -1,10 +1,13 @@
 import clang.cindex
 import inspect
 import importlib.util
+import ast
+import astunparse
 from typing import Optional, Callable
-from scripts.common import LLVM, LIBCLANG
 
-clang.cindex.Config.set_library_file(LIBCLANG)
+# from scripts.common import LLVM, LIBCLANG
+
+# clang.cindex.Config.set_library_file(LIBCLANG)
 
 
 def clang_get_func_code(
@@ -58,8 +61,46 @@ def clang_get_func_code_demangled(file_path: str, function_name: str):
     return clang_get_func_code(file_path, function_name, lambda node: node.displayname)
 
 
+class FunctionVisitor(ast.NodeVisitor):
+    def __init__(self, function_name):
+        self.function_name = function_name
+        self.functions = {}
+
+    def visit_FunctionDef(self, node):
+        # If this is the function we're looking for, or it's a function we've previously found
+        # and saved because it was called in the function we're looking for, save its source code.
+        if node.name == self.function_name or node.name in self.functions:
+            self.functions[node.name] = astunparse.unparse(node)
+
+        # If this function calls another function that we haven't found yet, save that function's name.
+        # The next time we encounter a FunctionDef node with that name, we'll save its source code.
+        for call in ast.walk(node):
+            if isinstance(call, ast.Call) and isinstance(call.func, ast.Name):
+                if call.func.id not in self.functions:
+                    self.functions[call.func.id] = None
+
+
 # todo: Java, Python
 def inspect_get_func_code_demangled(file_path: str, function_name: str):
+    source_code = ""
+    with open(file_path, "r") as source:
+        source_code = source_code.join(source.read())
+
+    tree = ast.parse(source_code)
+
+    visitor = FunctionVisitor(function_name)
+    visitor.visit(tree)
+
+    function_source_code = ""
+    for function_name, function_code in visitor.functions.items():
+        if function_code is not None:
+            function_source_code = function_source_code.join(function_code)
+
+    return function_source_code
+
+
+# Only get single function source code, can not deal with function call
+def inspect_get_func_code_demangled_2(file_path: str, function_name: str):
     spec = importlib.util.spec_from_file_location("module.name", file_path)
     module = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(module)
